buffer: test both implementations
