+++ go/vt/vtgate/buffer/buffer_keyspace.go
@@ -29,24 +29,15 @@ package buffer
 import (
 	"context"
 	"sync"
-	"time"
 
 	"vitess.io/vitess/go/sync2"
 	"vitess.io/vitess/go/vt/discovery"
-	"vitess.io/vitess/go/vt/log"
 	"vitess.io/vitess/go/vt/topo/topoproto"
 )
 
 type KeyspaceEventBuffer struct {
 	// Immutable configuration fields.
-	// Except for "now", they are parsed from command line flags.
-	// keyspaces has the same purpose as "shards" but applies to a whole keyspace.
-	keyspaces map[string]bool
-	// shards is a set of keyspace/shard entries to which buffering is limited.
-	// If empty (and *enabled==true), buffering is enabled for all shards.
-	shards map[string]bool
-	// now returns the current time. Overridden in tests.
-	now func() time.Time
+	config *Config
 
 	// bufferSizeSema limits how many requests can be buffered
 	// ("-buffer_size") and is shared by all shardBuffer instances.
@@ -68,83 +59,12 @@ type KeyspaceEventBuffer struct {
 }
 
 // New creates a new Buffer object.
-func New2() *KeyspaceEventBuffer {
-	return newWithNow2(time.Now)
-}
-
-func newWithNow2(now func() time.Time) *KeyspaceEventBuffer {
-	if err := verifyFlags(); err != nil {
-		log.Fatalf("Invalid buffer configuration: %v", err)
-	}
-	bufferSize.Set(int64(*size))
-	keyspaces, shards := keyspaceShardsToSets(*shards)
-
-	if *enabledDryRun {
-		log.Infof("vtgate buffer in dry-run mode enabled for all requests. Dry-run bufferings will log failovers but not buffer requests.")
-	}
-
-	if *enabled {
-		log.Infof("vtgate buffer enabled. PRIMARY requests will be buffered during detected failovers.")
-
-		// Log a second line if it's only enabled for some keyspaces or shards.
-		header := "Buffering limited to configured "
-		limited := ""
-		if len(keyspaces) > 0 {
-			limited += "keyspaces: " + setToString(keyspaces)
-		}
-		if len(shards) > 0 {
-			if limited == "" {
-				limited += " and "
-			}
-			limited += "shards: " + setToString(shards)
-		}
-		if limited != "" {
-			limited = header + limited
-			dryRunOverride := ""
-			if *enabledDryRun {
-				dryRunOverride = " Dry-run mode is overridden for these entries and actual buffering will take place."
-			}
-			log.Infof("%v.%v", limited, dryRunOverride)
-		}
-	}
-
-	if !*enabledDryRun && !*enabled {
-		log.Infof("vtgate buffer not enabled.")
-	}
-
+func NewKeyspaceEventBuffer(cfg *Config) *KeyspaceEventBuffer {
 	return &KeyspaceEventBuffer{
-		keyspaces:      keyspaces,
-		shards:         shards,
-		now:            now,
-		bufferSizeSema: sync2.NewSemaphore(*size, 0),
+		config:         cfg,
+		bufferSizeSema: sync2.NewSemaphore(cfg.Size, 0),
 		buffers:        make(map[string]*shardBufferKS),
 	}
-}
-
-// mode determines for the given keyspace and shard if buffering, dry-run
-// buffering or no buffering at all should be enabled.
-func (b *KeyspaceEventBuffer) mode(keyspace, shard string) bufferMode {
-	// Actual buffering is enabled if
-	// a) no keyspaces and shards were listed in particular,
-	if *enabled && len(b.keyspaces) == 0 && len(b.shards) == 0 {
-		// No explicit whitelist given i.e. all shards should be buffered.
-		return bufferEnabled
-	}
-	// b) or this keyspace is listed,
-	if b.keyspaces[keyspace] {
-		return bufferEnabled
-	}
-	// c) or this shard is listed.
-	keyspaceShard := topoproto.KeyspaceShardString(keyspace, shard)
-	if b.shards[keyspaceShard] {
-		return bufferEnabled
-	}
-
-	if *enabledDryRun {
-		return bufferDryRun
-	}
-
-	return bufferDisabled
 }
 
 // WaitForFailoverEnd blocks until a pending buffering due to a failover for
@@ -179,7 +99,7 @@ func (b *KeyspaceEventBuffer) HandleKeyspaceEvent(ksevent *discovery.KeyspaceEve
 	for _, shard := range ksevent.Shards {
 		sb := b.getOrCreateBuffer(shard.Target.Keyspace, shard.Target.Shard)
 		if sb != nil {
-			sb.recordKeyspaceEvent(shard.Serving, shard.LastReparenting)
+			sb.recordKeyspaceEvent(shard.Tablet, shard.Serving, shard.LastReparenting)
 		}
 	}
 }
@@ -205,7 +125,7 @@ func (b *KeyspaceEventBuffer) getOrCreateBuffer(keyspace, shard string) *shardBu
 	// Look it up again because it could have been created in the meantime.
 	sb, ok = b.buffers[key]
 	if !ok {
-		sb = newShardBufferKeyspace(b, b.mode(keyspace, shard), keyspace, shard)
+		sb = newShardBufferKeyspace(b, b.config.bufferingMode(keyspace, shard), keyspace, shard)
 		b.buffers[key] = sb
 	}
 	return sb
