+++ go/vt/vtgate/buffer/flags_test.go
@@ -23,6 +23,17 @@ import (
 )
 
 func TestVerifyFlags(t *testing.T) {
+	resetFlagsForTesting := func() {
+		// Set all flags to their default value.
+		flag.Set("enable_buffer", "false")
+		flag.Set("enable_buffer_dry_run", "false")
+		flag.Set("buffer_size", "10")
+		flag.Set("buffer_window", "10s")
+		flag.Set("buffer_keyspace_shards", "")
+		flag.Set("buffer_max_failover_duration", "20s")
+		flag.Set("buffer_min_time_between_failovers", "1m")
+	}
+
 	// Verify that the non-allowed (non-trivial) flag combinations are caught.
 	defer resetFlagsForTesting()
