+++ go/vt/vtgate/buffer/shard_buffer_keyspace.go
@@ -25,6 +25,7 @@ import (
 
 	"vitess.io/vitess/go/vt/log"
 	"vitess.io/vitess/go/vt/logutil"
+	"vitess.io/vitess/go/vt/proto/topodata"
 	"vitess.io/vitess/go/vt/topo/topoproto"
 	"vitess.io/vitess/go/vt/vterrors"
 )
@@ -58,6 +59,7 @@ type shardBufferKS struct {
 	queue []*entry
 	// lastStart is the last time we saw the start of a failover.
 	lastStart, lastEnd, lastReparent time.Time
+	currentPrimary                   *topodata.TabletAlias
 	// timeoutThread will be set while a failover is in progress and the object is in the BUFFERING state.
 	timeoutThread *timeoutThread
 	// wg tracks all pending Go routines. waitForShutdown() will use this field to
@@ -79,6 +81,10 @@ func newShardBufferKeyspace(buf *KeyspaceEventBuffer, mode bufferMode, keyspace,
 		logTooRecent:   logutil.NewThrottledLogger(fmt.Sprintf("FailoverTooRecent-%v", topoproto.KeyspaceShardString(keyspace, shard)), 5*time.Second),
 		state:          stateIdle,
 	}
+}
+
+func (sb *shardBufferKS) timeNow() time.Time {
+	return sb.buf.config.now()
 }
 
 // disabled returns true if neither buffering nor the dry-run mode is enabled.
@@ -115,13 +121,14 @@ func (sb *shardBufferKS) waitForFailoverEnd(ctx context.Context, keyspace, shard
 		// a) buffering was stopped recently
 		// OR
 		// b) we did not buffer, but observed a reparent very recently
-		now := sb.buf.now()
+		now := sb.timeNow()
+		minTimeBetweenFailovers := sb.buf.config.MinTimeBetweenFailovers
 
 		// a) Buffering was stopped recently.
 		// This can happen when we stop buffering while MySQL is not ready yet
 		// (read-only mode is not cleared yet on the new primary).
 		lastBufferingStopped := now.Sub(sb.lastEnd)
-		if !sb.lastEnd.IsZero() && lastBufferingStopped < *minTimeBetweenFailovers {
+		if !sb.lastEnd.IsZero() && lastBufferingStopped < minTimeBetweenFailovers {
 			sb.mu.Unlock()
 			msg := "NOT starting buffering"
 			if sb.mode == bufferDryRun {
@@ -130,7 +137,7 @@ func (sb *shardBufferKS) waitForFailoverEnd(ctx context.Context, keyspace, shard
 
 			sb.logTooRecent.Infof("%v for shard: %s because the last failover which triggered buffering is too recent (%v < %v)."+
 				" (A failover was detected by this seen error: %v.)",
-				msg, topoproto.KeyspaceShardString(keyspace, shard), lastBufferingStopped, *minTimeBetweenFailovers, err)
+				msg, topoproto.KeyspaceShardString(keyspace, shard), lastBufferingStopped, minTimeBetweenFailovers, err)
 
 			statsKeyWithReason := append(sb.statsKey, string(skippedLastFailoverTooRecent))
 			requestsSkipped.Add(statsKeyWithReason, 1)
@@ -143,7 +150,7 @@ func (sb *shardBufferKS) waitForFailoverEnd(ctx context.Context, keyspace, shard
 		// very low. If we do not skip buffering here, we would start buffering but
 		// not stop because we already observed the promotion of the new primary.
 		lastReparentAgo := now.Sub(sb.lastReparent)
-		if !sb.lastReparent.IsZero() && lastReparentAgo < *minTimeBetweenFailovers {
+		if !sb.lastReparent.IsZero() && lastReparentAgo < minTimeBetweenFailovers {
 			sb.mu.Unlock()
 			msg := "NOT starting buffering"
 			if sb.mode == bufferDryRun {
@@ -152,7 +159,7 @@ func (sb *shardBufferKS) waitForFailoverEnd(ctx context.Context, keyspace, shard
 
 			sb.logTooRecent.Infof("%v for shard: %s because the last reparent is too recent (%v < %v)."+
 				" (A failover was detected by this seen error: %v.)",
-				msg, topoproto.KeyspaceShardString(keyspace, shard), lastReparentAgo, *minTimeBetweenFailovers, err)
+				msg, topoproto.KeyspaceShardString(keyspace, shard), lastReparentAgo, minTimeBetweenFailovers, err)
 
 			statsKeyWithReason := append(sb.statsKey, string(skippedLastReparentTooRecent))
 			requestsSkipped.Add(statsKeyWithReason, 1)
@@ -209,12 +216,12 @@ func (sb *shardBufferKS) startBufferingLocked(err error) {
 	lastRequestsDryRunMax.Set(sb.statsKey, 0)
 	failoverDurationSumMs.Reset(sb.statsKey)
 
-	sb.lastStart = sb.buf.now()
+	sb.lastStart = sb.timeNow()
 	sb.logErrorIfStateNotLocked(stateIdle)
 	sb.state = stateBuffering
 	sb.queue = make([]*entry, 0)
 
-	sb.timeoutThread = newTimeoutThread(sb)
+	sb.timeoutThread = newTimeoutThread(sb, sb.buf.config.MaxFailoverDuration)
 	sb.timeoutThread.start()
 	msg := "Starting buffering"
 	if sb.mode == bufferDryRun {
@@ -222,7 +229,13 @@ func (sb *shardBufferKS) startBufferingLocked(err error) {
 	}
 	starts.Add(sb.statsKey, 1)
 	log.Infof("%v for shard: %s (window: %v, size: %v, max failover duration: %v) (A failover was detected by this seen error: %v.)",
-		msg, topoproto.KeyspaceShardString(sb.keyspace, sb.shard), *window, *size, *maxFailoverDuration, err)
+		msg,
+		topoproto.KeyspaceShardString(sb.keyspace, sb.shard),
+		sb.buf.config.Window,
+		sb.buf.config.Size,
+		sb.buf.config.MaxFailoverDuration,
+		err,
+	)
 }
 
 // logErrorIfStateNotLocked logs an error if the current state is not "state".
@@ -269,7 +282,7 @@ func (sb *shardBufferKS) bufferRequestLocked(ctx context.Context) (*entry, error
 
 	e := &entry{
 		done:     make(chan struct{}),
-		deadline: sb.buf.now().Add(*window),
+		deadline: sb.timeNow().Add(sb.buf.config.Window),
 	}
 	e.bufferCtx, e.bufferCancel = context.WithCancel(ctx)
 	sb.queue = append(sb.queue, e)
@@ -412,12 +425,13 @@ func (sb *shardBufferKS) remove(toRemove *entry) {
 	// Entry was already removed. Keep the queue as it is.
 }
 
-func (sb *shardBufferKS) recordKeyspaceEvent(shardStillServing bool, lastReparentForShard time.Time) {
+func (sb *shardBufferKS) recordKeyspaceEvent(currentPrimary *topodata.TabletAlias, shardStillServing bool, lastReparentForShard time.Time) {
 	sb.mu.Lock()
 	defer sb.mu.Unlock()
 
 	log.Infof("disruption in shard %s/%s resolved (serving: %v)", sb.keyspace, sb.shard, shardStillServing)
 
+	sb.currentPrimary = currentPrimary
 	sb.lastReparent = lastReparentForShard
 	if shardStillServing {
 		sb.stopBufferingLocked(stopFailoverEndDetected, "a primary promotion has been detected")
@@ -431,7 +445,7 @@ func (sb *shardBufferKS) stopBufferingDueToMaxDuration() {
 	defer sb.mu.Unlock()
 
 	sb.stopBufferingLocked(stopMaxFailoverDurationExceeded,
-		fmt.Sprintf("stopping buffering because failover did not finish in time (%v)", *maxFailoverDuration))
+		fmt.Sprintf("stopping buffering because failover did not finish in time (%v)", sb.buf.config.MaxFailoverDuration))
 }
 
 func (sb *shardBufferKS) stopBufferingLocked(reason stopReason, details string) {
@@ -440,7 +454,7 @@ func (sb *shardBufferKS) stopBufferingLocked(reason stopReason, details string) 
 	}
 
 	// Stop buffering.
-	sb.lastEnd = sb.buf.now()
+	sb.lastEnd = sb.timeNow()
 	d := sb.lastEnd.Sub(sb.lastStart)
 
 	statsKeyWithReason := append(sb.statsKey, string(reason))
@@ -450,11 +464,11 @@ func (sb *shardBufferKS) stopBufferingLocked(reason stopReason, details string) 
 	failoverDurationSumMs.Add(sb.statsKey, int64(d/time.Millisecond))
 	if sb.mode == bufferDryRun {
 		utilDryRunMax := int64(
-			float64(lastRequestsDryRunMax.Counts()[sb.statsKeyJoined]) / float64(*size) * 100.0)
+			float64(lastRequestsDryRunMax.Counts()[sb.statsKeyJoined]) / float64(sb.buf.config.Size) * 100.0)
 		utilizationDryRunSum.Add(sb.statsKey, utilDryRunMax)
 	} else {
 		utilMax := int64(
-			float64(lastRequestsInFlightMax.Counts()[sb.statsKeyJoined]) / float64(*size) * 100.0)
+			float64(lastRequestsInFlightMax.Counts()[sb.statsKeyJoined]) / float64(sb.buf.config.Size) * 100.0)
 		utilizationSum.Add(sb.statsKey, utilMax)
 	}
 
@@ -488,12 +502,12 @@ func (sb *shardBufferKS) drain(q []*entry, err error) {
 	// shardBuffer as well e.g. to get the current oldest entry.
 	sb.timeoutThread.stop()
 
-	start := sb.buf.now()
+	start := sb.timeNow()
 	// TODO(mberlin): Parallelize the drain by pumping the data through a channel.
 	for _, e := range q {
 		sb.unblockAndWait(e, err, true /* releaseSlot */, true /* blockingWait */)
 	}
-	d := sb.buf.now().Sub(start)
+	d := sb.timeNow().Sub(start)
 	log.Infof("Draining finished for shard: %s Took: %v for: %d requests.", topoproto.KeyspaceShardString(sb.keyspace, sb.shard), d, len(q))
 	requestsDrained.Add(sb.statsKey, int64(len(q)))
 
@@ -518,14 +532,14 @@ func (sb *shardBufferKS) waitForShutdown() {
 // sizeForTesting is used by the unit test only to find out the current number
 // of buffered requests.
 // TODO(mberlin): Remove this if we add a more general statistics reporting.
-func (sb *shardBufferKS) sizeForTesting() int {
+func (sb *shardBufferKS) testGetSize() int {
 	sb.mu.RLock()
 	defer sb.mu.RUnlock()
 	return len(sb.queue)
 }
 
 // stateForTesting is used by unit tests only to probe the current state.
-func (sb *shardBufferKS) stateForTesting() bufferState {
+func (sb *shardBufferKS) testGetState() bufferState {
 	sb.mu.RLock()
 	defer sb.mu.RUnlock()
 	return sb.state
